feat: Extract ButtonType from CrossLoginUI & MyKSuite to its own shared location
