+++ Compose/Basics/build.gradle.kts
@@ -42,5 +42,6 @@ dependencies {
     implementation(platform(core.compose.bom))
     implementation(core.compose.runtime)
     implementation(core.compose.ui.tooling)
+    implementation(core.foundation.android)
     implementation(core.splitties.collections)
 }
