+++ CrossAppLoginUI/src/main/kotlin/com/infomaniak/core/crossloginui/views/CrossLoginListAccountsView.kt
@@ -27,9 +27,9 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.AbstractComposeView
+import com.infomaniak.core.compose.basics.ButtonType
 import com.infomaniak.core.compose.materialthemefromxml.MaterialThemeFromXml
 import com.infomaniak.core.crossloginui.R
-import com.infomaniak.core.crossloginui.data.CrossLoginButtonType
 import com.infomaniak.core.crossloginui.data.CrossLoginDefaults
 import com.infomaniak.core.crossloginui.data.CrossLoginUiAccount
 import com.infomaniak.core.crossloginui.utils.getColorOrNull
@@ -55,7 +55,7 @@ class CrossLoginListAccountsView @JvmOverloads constructor(
     private var descriptionColor: Color?
     private var avatarStrokeColor: Color?
     private var buttonStrokeColor: Color?
-    private var buttonType: CrossLoginButtonType?
+    private var buttonType: ButtonType?
 
     init {
         context.obtainStyledAttributes(attrs, R.styleable.CrossLoginListAccountsView, defStyleAttr, 0).apply {
@@ -66,7 +66,7 @@ class CrossLoginListAccountsView @JvmOverloads constructor(
             avatarStrokeColor = getColorOrNull(R.styleable.CrossLoginListAccountsView_crossLoginAvatarStrokeColor)
             buttonStrokeColor = getColorOrNull(R.styleable.CrossLoginListAccountsView_crossLoginButtonStrokeColor)
             val crossLoginButtonType = getStringOrNull(R.styleable.CrossLoginListAccountsView_crossLoginButtonType)
-            buttonType = enumValueOfOrNull<CrossLoginButtonType>(crossLoginButtonType)
+            buttonType = enumValueOfOrNull<ButtonType>(crossLoginButtonType)
 
             recycle()
         }
