+++ MyKSuite/build.gradle.kts
@@ -50,6 +50,7 @@ dependencies {
 
     implementation(project(":Core"))
     implementation(project(":Core:Avatar"))
+    implementation(project(":Core:Compose:Basics"))
     implementation(project(":Core:Compose:Margin"))
     implementation(project(":Core:Compose:MaterialThemeFromXml"))
