+++ MyKSuite/src/main/java/com/infomaniak/core/myksuite/ui/components/MyKSuitePrimaryButton.kt
@@ -30,8 +30,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.tooling.preview.Preview
+import com.infomaniak.core.compose.basics.ButtonType
 import com.infomaniak.core.compose.margin.Margin
-import com.infomaniak.core.myksuite.ui.screens.components.MyKSuiteButtonType
 import com.infomaniak.core.myksuite.ui.theme.Dimens
 import com.infomaniak.core.myksuite.ui.theme.MyKSuiteTheme
 import com.infomaniak.core.myksuite.ui.theme.Typography
@@ -65,12 +65,12 @@ private fun Preview() {
             Column(verticalArrangement = Arrangement.spacedBy(Margin.Large)) {
                 MyKSuitePrimaryButton(
                     text = "Lorem",
-                    shape = MyKSuiteButtonType.Mail.shape,
+                    shape = ButtonType.Mail.shape,
                     onClick = {},
                 )
                 MyKSuitePrimaryButton(
                     text = "Close",
-                    shape = MyKSuiteButtonType.Drive.shape,
+                    shape = ButtonType.Drive.shape,
                     onClick = {},
                 )
             }
